backend/app/connectors/hms_connector.py
import mysql.connector
from .base import BaseConnector, DiagnosticTask
from typing import List, Dict
import os
import logging

logger = logging.getLogger("HMSConnector")

class HMSConnector(BaseConnector):
    def __init__(self):
        self.conn = None
        self.db_config = {
            "host": "localhost",
            "user": "root",
            "password": "",  # To be configured
            "database": "hms_suite"
        }
    
    def connect(self):
        try:
            # self.conn = mysql.connector.connect(**self.db_config)
            # logger.info("Connected to HMS Database")
            pass
        except Exception as e:
            logger.error(f"Failed to connect to DB: {e}")

    def fetch_pending_tasks(self) -> List[DiagnosticTask]:
        """
        Polls the `patient_xray_report_details` table.
        Assumes columns: ID, Patient_id, File_Path (or similar).
        """
        tasks = []
        if not self.conn:
             # MOCK DATA FOR DEMO
            return [
                DiagnosticTask(
                    id="101", 
                    patient_id="P-505", 
                    file_path="C:/hospital_data/patient_505_chest_xray.jpg", 
                    file_type="XRAY"
                )
            ]
        
        cursor = self.conn.cursor(dictionary=True)
        # Query logic (Adjust column names based on actual schema)
        # query = "SELECT * FROM patient_xray_report_details WHERE status = 'PENDING' LIMIT 5"
        # cursor.execute(query)
        # for row in cursor.fetchall():
        #     tasks.append(DiagnosticTask(
        #         id=str(row['ID']),
        #         patient_id=str(row['Patient_id']),
        #         file_path=row.get('File_Path', ''),
        #         file_type='XRAY'
        #     ))
        return tasks

    def update_result(self, task_id: str, findings: Dict):
        logger.info(f"Updating Task {task_id} with findings: {findings}")
        if self.conn:
            # cursor = self.conn.cursor()
            # sql = "UPDATE patient_xray_report_details SET machine_findings = %s, status = 'COMPLETED' WHERE ID = %s"
            # cursor.execute(sql, (str(findings), task_id))
            # self.conn.commit()
            pass
